pyproject.toml
[tool.poetry]
name = "testing"
version = "0.1.0"
description = ""
authors = ["drforse <george.lifeslice@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.13.1"
mysqlclient = "^2.1.0"
redis = "^4.1.1"
celery = "^5.2.3"
setuptools = "59.7.0"
gunicorn = "^20.1.0"
drf-yasg = {extras = ["validation"], version = "^1.20.0"}
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
flower = "^1.0.0"
django-stubs = "^1.9.0"
mypy = "^0.931"
celery-stubs = "^0.1.2"
pip = "^21.3.1"
install = "^1.3.5"
djangorestframework-stubs = "^1.4.0"
pytest-django = "^4.5.2"
pytest-celery = "^0.0.0"
eventlet = "^0.33.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
